removed history.base + other clean ups
