+++ docker-compose.yml
@@ -9,6 +9,7 @@ services:
       - postgres
     volumes: 
       - ".:/usr/src/app"
+    command: python manage.py runserver 0.0.0.0:8000 --settings=settings.development
 
   postgres:
     container_name: hocg-postgres
