+++ history/historyofcg/uwsgi.py
@@ -2,8 +2,6 @@ import os
 import sys
 import site
 
-site.addsitedir('/opt/apps/env/lib/python2.6/site-packages')
-
 os.environ['DJANGO_SETTINGS_MODULE'] = 'historyofcg.settings'
 
 import django.core.handlers.wsgi
